requirements.txt
sentence-transformers
chromadb  
pdfplumber
python-docx
tiktoken